+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changes for lovely-db-testing
 
+## unreleased
+
+- allow to configure server container in setup
+- allow to null the test and dev directory in pg setup
+
 ## 2021-10-19 / 0.0.1
 
 - initial junit extension implemented